Create a deep copy of existing `fileHistoryUi`
